requirements.txt
appnope==0.1.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
decorator==4.4.1
Django==2.2.7
djangorestframework==3.10.3
importlib-metadata==0.23
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.0
packaging==19.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
Pygments==2.5.1
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
